web_ui/requirements.txt
Flask==2.3.2
python-chess==1.10.0
